+++ stories/Autocomplete.stories.js
@@ -59,7 +59,7 @@ storiesOf('Autocomplete', module)
   .add('Enter mnemonics (9-word mnemonic) - not sorted', () => (
     <Autocomplete
       label="Recovery phrase"
-      suggestedWords = {OPTIONS}
+      options = {OPTIONS}
       placeholder="Enter mnemonic..."
       sortAlphabetically={false}
       multipleSameSelections={false}
@@ -71,7 +71,7 @@ storiesOf('Autocomplete', module)
   .add('Enter mnemonics (9-word mnemonic) - sorted with multiple same selections', () => (
     <Autocomplete
       label="Recovery phrase"
-      suggestedWords = {OPTIONS}
+      options = {OPTIONS}
       placeholder="Enter mnemonic..."
       maxSelections={9}
       skin={<SimpleAutocompleteSkin />}
@@ -81,10 +81,10 @@ storiesOf('Autocomplete', module)
   .add('Enter mnemonics - (12-word mnemonic) with 5 visible suggestions', () => (
     <Autocomplete
       label="Recovery phrase"
-      suggestedWords = {OPTIONS}
+      options = {OPTIONS}
       placeholder="Enter mnemonic..."
       maxSelections={12}
-      maxVisibleSuggestions={5}
+      maxVisibleOptions={5}
       skin={<SimpleAutocompleteSkin />}
     />
   ))
@@ -92,10 +92,10 @@ storiesOf('Autocomplete', module)
   .add('Enter mnemonics - (12-word mnemonic) with 5 visible suggestions and regex that allows only letters', () => (
     <Autocomplete
       label="Recovery phrase"
-      suggestedWords = {OPTIONS}
+      options = {OPTIONS}
       placeholder="Enter mnemonic..."
       maxSelections={12}
-      maxVisibleSuggestions={5}
+      maxVisibleOptions={5}
       invalidCharsRegex= {/[^a-zA-Z]/g}
       skin={<SimpleAutocompleteSkin />}
     />
@@ -115,9 +115,9 @@ storiesOf('Autocomplete', module)
           <Autocomplete
             label="Recovery phrase"
             placeholder="Enter recovery phrase"
-            suggestedWords = {OPTIONS}
+            options = {OPTIONS}
             maxSelections={12}
-            maxVisibleSuggestions={5}
+            maxVisibleOptions={5}
             invalidCharsRegex= {/[^a-zA-Z]/g}
             skin={<SimpleAutocompleteSkin />}
           />